requirements.txt
langchain
google-api-python-client
google-auth
streamlit
python-dotenv
fastapi
uvicorn
requests
langchain_community
google-generativeai


